+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 + Make console color default disabled
++ Allow concurrent tests to work using associated objects instead of global state
 
 ## 0.2
 + Add support for native feature files